Working on speed calculation
Speed stopped working quickly with use (zero or inf), although worked fine off the car.  Just in case, speed calculation was taken out of the interrupt to speed up the calculation, and time interval will be shown on the display to see if it drops to zero.  PLUS, in case of signal bounce, a condition of time of at least 10 ms is included (corresponding to about 100 m/s!).