Sample data logger with backup
In order to keep an accurate record of battery charge, the arduino needs to be able to recover the last data after a power interuption.  This program shows how this problem can be solved.  There is a lot of data storage and conversion, working with character strings, etc.
Comments at the top.